requirements.txt
pycryptodome
pymongo
tornado
tqdm
lorem
python-decouple
coverage